chore(starknet_integration_tests): make dump fn part of executable setup (#4294)
* chore(starknet_integration_tests): include config in execution setup

commit-id:3e7c2717

* refactor(starknet_integration_tests): consolidate config utils into executable setup

commit-id:21bf81ea

* chore(starknet_integration_tests): change arg to json dump fn

commit-id:dd9072d0

* chore(starknet_integration_tests): make dump fn part of executable setup

commit-id:1067873a
