+++ crates/starknet_integration_tests/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod config_utils;
 pub mod flow_test_setup;
 pub mod integration_test_setup;
 pub mod integration_test_utils;
